refactor(ui): replace css with scss
