+++ ui/package.json
@@ -45,6 +45,7 @@
     "jest-environment-jsdom": "^28.1.1",
     "prettier": "^2.6.2",
     "react-test-renderer": "^18.1.0",
+    "sass": "^1.53.0",
     "ts-jest": "^28.0.5",
     "typescript": "^4.6.3",
     "vite": "^2.9.9"
