+++ ui/src/components/Button/index.tsx
@@ -1,5 +1,5 @@
 import {ReactNode} from 'react';
-import styles from './styles.module.css';
+import styles from './styles.module.scss';
 
 interface ButtonProps {
   className?: string;
