+++ ui/src/components/ClubsCarousel/index.tsx
@@ -1,4 +1,4 @@
-import styles from './styles.module.css';
+import styles from './styles.module.scss';
 import Carousel from 'react-multi-carousel';
 import 'react-multi-carousel/lib/styles.css';
 import {dummyClubLogo} from '../../assets';
