+++ ui/src/components/EventsCarousel/styles.module.scss
