+++ ui/src/components/FormComponents/Input.tsx
@@ -4,7 +4,7 @@ import {
   FieldValues,
   UseControllerProps,
 } from 'react-hook-form';
-import styles from './styles.module.css';
+import styles from './styles.module.scss';
 
 interface Props<T> extends UseControllerProps<T> {
   label: string;
