+++ ui/src/components/NewsCarousel/styles.module.scss
@@ -58,7 +58,8 @@
   border-radius: 0 32px 32px 0;
 }
 
-@media only screen and (max-width: 700px) {
+@import '../../index';
+@media only screen and (max-width: $sm) {
   .rightImg {
     display: none !important;
   }
