+++ ui/src/components/ProfileCard/styles.module.scss
@@ -42,7 +42,8 @@
   right: 10vh;
 }
 
-@media only screen and (max-width: 700px) {
+@import '../../index';
+@media only screen and (max-width: $sm) {
   .cardContainer {
     flex-direction: column;
   }
