+++ ui/src/components/Sidebar/SidebarRow.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './styles.module.css';
+import styles from './styles.module.scss';
 
 interface Props {
   icon: string;
