+++ ui/src/components/Sidebar/styles.module.scss
