+++ ui/src/components/TopBar/index.tsx
@@ -1,4 +1,4 @@
-import styles from './styles.module.css';
+import styles from './styles.module.scss';
 
 export const TopBar: React.FC = () => {
   return (
