+++ ui/src/index.scss
@@ -12,6 +12,8 @@ body {
   box-sizing: border-box;
 }
 
+$sm: 700px;
+
 :root {
   --primary-bg: white;
   --secondary-bg: #829bb0;
