+++ ui/src/pages/Login/styles.module.scss
@@ -63,7 +63,8 @@
   border-radius: 0 0 40px 40px;
 }
 
-@media only screen and (max-width: 700px) {
+@import '../../index';
+@media only screen and (max-width: $sm) {
   .container {
     background-size: auto;
   }
