+++ ui/src/pages/Profile/styles.module.scss
@@ -12,7 +12,8 @@
   width: 82%;
 }
 
-@media only screen and (max-width: 700px) {
+@import '../../index';
+@media only screen and (max-width: $sm) {
   .NavContainer {
     display: none;
   }
